Move all logic into analysis page rather than doing calculations on home page
